add anlysis models file
